requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
requests==2.9.1
Pillow==3.1.1
django-cities==0.4.2
django-registration==2.0.4
django-haystack==2.4.1
django-tinymce==2.2.0
django-paypal==0.3
django-modeltranslation==0.11
